Updated jnbt to 1.2 Beta Fixed regions bleeding through to all worlds (This should fix a lot of the problems seen in 5.9.0) Added /r list [owned/playername] as a command Removed /r list-owned command Made GamemodeCacheManager and InventoryCacheManager local to the indiviual regions instead of global to fix issue with regions within regions not switching properly Updated Painting events to new Hanging events Cleaned up some code in RegiosAPI and Regios Fixed /r expand max command throwing an NPE when creating a poly region

+++ src/net/jzx7/regios/Commands/MiscCommands.java
@@ -192,8 +192,8 @@ public class MiscCommands extends PermissionsCore {
 				rv2 = ((CuboidRegion) r).getL2().toVector();
 			}
 			double x_mid, z_mid;
-			x_mid = (double)(Math.max(rv1.getX(), rv2.getX()) - Math.min(rv1.getX(), rv2.getX()));
-			z_mid = (double)(Math.max(rv1.getZ(), rv2.getZ()) - Math.min(rv1.getZ(), rv2.getZ()));
+			x_mid = (Math.max(rv1.getX(), rv2.getX()) - Math.min(rv1.getX(), rv2.getX()));
+			z_mid = (Math.max(rv1.getZ(), rv2.getZ()) - Math.min(rv1.getZ(), rv2.getZ()));
 			double hypot_x, hypot_z;
 			hypot_x = origin.getX()- x_mid;
 			hypot_z = origin.getZ() - z_mid;
