+++ src/net/jzx7/regios/RBF/RBF_Load.java
@@ -121,7 +121,7 @@ public class RBF_Load extends PermissionsCore {
 						ih.getInventory().clear();
 					}
 
-					b.setTypeId((int) blocks[index]);
+					b.setTypeId(blocks[index]);
 					b.setData(blockData[index]);
 
 					if(b.getType().equals(Material.CHEST)) { //Added to load chest/furnace/etc inventory from backup
